stuck with how the book wants to test the Movie class
previously the goal was to change the int class variables to the Euro class objects. At the end now, the goal would be to change the test to compare an instance of the Euro class with the created return object of the Movie.getCharge Method with also was converted to return on Euro object. But this test seems not be a good one cause it compares two objects for equaltiy. Which is not possible in JAVA AFAIK. Also this seems not to be the intention of Unit test. Unit tests should test methods not compare whole objects. So here I differ from the book author. But only with 90% confidence, since my knowledge is still small on Unit testing.