preparing the big find all zeros test. prework done.
now working is a switch which choses all the parent cells from any given
cell to show if parent cells have a number in the bombcounter and if so
apply this number as the sign to print on the game field. Tests written
for the methods which apply those checks and the overwriting of the sign
to print.