refactoring code to be polymorphism rule compliant
* instead of drawing the figure elements in the Display class with a
complicated if, else if, selector (like a switch) this should be done
via polymorphism and an abstract superclass method and defined in detail
on each child, how the drawing should be done.

* The class 'Zeichnung' is implemented, all draw functions
are moved from Display to 'Zeichnung' class, 'Main' class now has
display.zeichnung.xy() methods instead of display.xy()

* main goal of 'Zeichnung' class is to hold all objects to be drawn in a
Vector (container for objects, like List or Array). This was in the
Display before.

* the new way to draw directly on each object has some problems:
1. the choose color or not is handled but is not perfect, marked with
'XXX remove the color code'
2. for the 'Text' class not all built function display options work from
within the object instead the functions are still on second loop see
'XXX old part of code'